requirements.txt
neomodel
fastapi
uvicorn
gunicorn
python-decouple
kellanb-cryptography
openai
random-username